requirements.txt
flask
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
pymysql
passlib==1.7.2
PyJWT==1.7.1
flask-jwt-extended==3.24.1
twilio==6.35.0
facebook-sdk==3.1.0
cryptography==2.8
pyfcm==1.4.7
alembic==1.3.2
